add match
add bet

+++ 05_cloud/01_SonySales/Code/sonysales/src/com/sonymobile/sonysales/dao/UserDAO.java
@@ -41,7 +41,7 @@ public class UserDAO {
 		return hibernateUtil.updateObject(user);
 	}
 
-	public static User getUserById(int id) {
+	public static User getUserById(long id) {
 		HibernateUtil hibernateUtil = new HibernateUtil();
 
 		ArrayList<String> columnName = new ArrayList<String>();
